correcção na estrutura do cartao da noticia, badge num notificaçoes (apenas no login)

+++ marianaisabel-backend-proj5/src/main/java/pt/uc/dei/proj5/rest/NotificationController.java
@@ -101,7 +101,7 @@ public class NotificationController {
 
 		try {
 			User user = userService.getNonDeletedEntityByToken(authString); // utilizador dono das notificações 
-			int resultado = notificationService.getNumberUnreadNotifFromUser(user);
+			Long resultado = notificationService.getNumberUnreadNotifFromUser(user);
 			return Response.ok(resultado).build();
 					
 		} catch (Exception e) {
